feat : is_admin SQL 오류 수정 1

+++ backend/src/main/java/com/codestates/common/delete/SchedulingConfig.java
@@ -22,6 +22,8 @@ public class SchedulingConfig {
         LocalDateTime now = LocalDateTime.now();
 
         List<Member> deleteMemberList = memberRepository.findByDeletionDateBefore(now);
-        memberRepository.deleteAll(deleteMemberList);
+        if (deleteMemberList != null) {
+            memberRepository.deleteAll(deleteMemberList);
+        }
     }
 }
